requirements.txt
numpy
PyQt5
pandas
pytesseract
pillow
opencv-python
